op-node: Fix retrying dialing at startup
I needed to move IsURLAvailable around to break an import cycle.
I have manually tested that the op-node now properly waits for URLs
to be up before erroring.
